api/requirements.txt
Flask==2.3.2
pymongo==4.5.0
PyYAML==6.0